requirements.txt
amqp==2.3.2
asn1crypto==0.24.0
billiard==3.5.0.3
celery==4.2.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
crayons==0.1.2
cryptography==2.2.2
Django==2.0.6
django-appconf==1.0.2
django-bootstrap-form==3.4
django-cors-headers==2.4.0
django-ipware==1.1.6
django-user-accounts==2.0.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.6
itypes==1.1.0
Jinja2==2.10
jsonfield==2.0.2
kombu==4.2.1
MarkupSafe==1.0
numpy==1.14.3
pandas==0.23.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyaml==17.12.1
pycparser==2.18
pycrypto==2.6.1
PyJWT==1.6.4
pyOpenSSL==18.0.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
requests==2.18.4
simple-salesforce==0.74.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
gunicorn==19.9.0
